Add a proper vite config for proxying to /api and then add some more form components that can upload images and stuff
