+++ frontend/package.json
@@ -22,6 +22,7 @@
     "@tanstack/react-query": "^5.51.3",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
+    "jszip": "^3.10.1",
     "lucide-react": "^0.380.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
